+++ pcs/output.py
@@ -1,13 +1,13 @@
 import tempfile
 from pathlib import Path
-from typing import Any, Dict, Optional
+from typing import Any, Optional
 
 from mlflow.utils.mlflow_tags import MLFLOW_RUN_NAME
 
 from pcs.command import Command
 from pcs.exceptions import PythonComponentSystemException
+from pcs.mlflow_run import MLflowRun
 from pcs.registry import Registry
-from pcs.run import MLflowRun
 
 
 class Output(MLflowRun):
@@ -17,12 +17,12 @@ class Output(MLflowRun):
     COMMAND_REGISTRY_FILENAME = "pcs.command_registry.yaml"
     """
     Output from running a Command.
-     
+
     As its name implies, an Output documents an instance of code execution
     and records output associated with it (similar to a logger). This is also
     called "tracking" by the MLflow project (and this class's parent is
     MLflowRun, which is a wrapper around the the MLflowRun abstraction).
-    
+
     In addition to tracking the output of a code that was run, an Output
     can also be used for reproducibility: an Output has a member called a
     `command` that can be used to perform a "re-run" of the the same function
@@ -48,7 +48,7 @@ class Output(MLflowRun):
         )
 
     @classmethod
-    def from_existing_mlflow_run(cls, run_id: str) -> "Run":
+    def from_existing_mlflow_run(cls, run_id: str) -> "Output":
         output = super().from_existing_mlflow_run(run_id)
         output._command = output._fetch_command()
         return output
